sqlgen: split dialect file
Organize runtime dialect functions into one source file per dialect
for better navigability. In preparation to add another function.
