+++ sqlgen/dialect_sqlite3.go
@@ -1,0 +1,13 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// Copyright (C) 2017 Space Monkey, Inc.
+// See LICENSE for copying information.
+
+package sqlgen
+
+// this type is specially named to match up with the name returned by the
+// dialect impl in the sql package.
+type sqlite3 struct{}
+
+func (s sqlite3) Rebind(sql string) string {
+	return sql
+}
