+++ sqlgen/doc.go
@@ -1,0 +1,7 @@
+// Copyright (C) 2023 Storj Labs, Inc.
+
+/*
+Package sqlgen contains code which is included as-is in the final template.
+Package sqlgen/sqlbundle joins all the code in package sqlgen to a single string.
+*/
+package sqlgen
